Modifiy Performance Counter for DSP Inference
This commit modifies the cs_registers module to allow inference of DSP
slices to absorb counter logic. A preprocessor variable is introduced
to include the required synthesis pragma. DSP inferrence is supported
for Xilinx FPGA devices featuring DSP48E1 slices or similar and for
counter widths up to 48 bit.

The resulting relative logic utilizations with respect to the non-DSP
32 bit counter implementation are 84% and 90% respectively for 32 and
48 bit counters. For 32 bit counters only, the corresponding
flip-flops can be incorporated into the DSP's output pipeline
register, resulting in a reduction of the number of flip-flops to 50%.
In order to infer DSP slices for performance counters, define the
preprocessor variable XILINX_DSP_COUNTERS.

According to discussion in PR, the actual counter implementations
have been moved to a separate module ibex_counters.sv

* ibex_counters.sv:
        Implements counter logic. The preprocessor variable
                FPGA_XILINX controls insertion of a DSP Pragma for
                supported xilinx devices.

        Parameters:

                * MaxNumCounters: Defines the number of supporting
                signals (counter_inc_i, counter_we_i, counter_val_o)
                signals connected to this module.

                * NumCounters: Defines the number of counters to be
                elaborated in this module. Counters
                [MaxNumCounters-1:NumCounters] are connected to '0.

                * CounterWidth: Counter bit-width.

* ibex_cs_registers.sv:
        Now instantiates three instances of the ibex_counter module.

        mcycle_counter_i (1/1), minstret_counter_i (1/1),
        mcounters_variable_i (MHPMCounterNum / 29).