Split have_instr definition out of always block in if_stage
This should cause no functional change, but avoids a seeming
combinatorial loop reported by Verilator.

The seeming loop is because the always_comb process that contained
have_instr is sensitive to if_id_pipe_reg_we but that process wrote
the have_instr signal, which is used in the continuous assignment to
if_id_pipe_reg_we later on.