[dv] Remove duplicate lowrisc:constants:top_pkg
The core file and SV package were copied in here, which breaks the
OpenTitan build. The constants file is provided there too, but a
recent builds will pick up this one here. This is incorrect.

As the core file seems to be unused anyways, remove it from here along
with the package.

Added the, as far as I can see, only used constant here for
simplicity, alternatively it is an option to properly depend on the
other core.

Not tested because of tool availability.