Verible Format: Do not format port list
The formatting of the port list with Verible isn't working very well
currently:
- It forces four instead of two space indent
- Comments are not aligned
- Multi-line comments aren't supported.

All these issues are tracked in
https://github.com/lowRISC/ibex/issues/687. For now, disable formatting
of port lists and rely on manually formatting those.

Note that this option doesn't affect parameter lists in the module
header, which are still formatted wrong in a similar way as ports, which
no option to disable that behavior.