Format source code with Verible, part 1
This commit is the first, incomplete run of Verible format on the source
code in the `rtl` directory. The goal is to get the "easy cases" out of
the way, the ones we agree on that the produced formatting is good
enough. This should leave us with the cases where we either don't agree
with the formatting produced by Verible format (either because it
violates the lowRISC style guide, or because it hurts readability), or
the cases where Verible format breaks down. Ultimately, we should be
able to reduce the diff between what verible-verilog-format produces,
and our code bases, iteratively to zero.

The commit doesn't contain all changes made by Verible format:
- All changes to module headers (parameters and port lists) have been
  excluded due to known formatting issues there.
- Many changes to case statements have been excluded, as there are a
  couple formatting issues and inconsistencies.
- Some assign statement changes have been excluded for too badly hurting
  readbility.
- Some other obvious misformattings have been corrected or excluded.

Most limitations of Verible format are tracked in
https://github.com/lowRISC/ibex/issues/687. Not all problems I've seen
have been filed upstream yet,

The diff contains a rather exhaustive list of changes which I
pre-filtered to remove the most obvious "ugly ones" (see above). That
still leaves a couple ones which are worth discussing:

- Verible doesn't use tabular style for blocks of signal definitions.
  Doing that would most likely be possible in an algorithmic way if we
  agree that readability is hurt too much by not doing it.
- Verible doesn't use tabular style for other assignment code. Doing
  that in a consistent, automated fashion is probably harder. Also, the
  readability trade-offs are different.

I invite you all to look at the diff and classify the changes:
- Changes you're OK with.
- Changes which might not be your preferred style, but are not too bad
  either.
- Changes which you think severely hurt readability, or are against the
  style guide.

I'll go through those comments and remove all contentious items from
the diff; we can then discuss them individually, also involving the
style guide team and the verible developers.

In general, the trade off is as follows: Verible format takes away the
arguing and manual formatting completely out of the code writing and
review process. We win a significant amount of time there. There will
always be cases where a manual formatting might be marginally better,
but expressing that into an automated formatting rule is (close to)
impossible. In those cases, we win by "swallowing the pill" of not fully
optimal styling, for the benefit of never having to think about styling
again (TM).