Add X-Interface UVM Agent
This adds an X-Interface UVM response agent. It emulates the bitmanip
extension (zbb and zbt) to cover example instructions for all use
cases of the interface currently supported by ibex.

Caution: This has been my first experience with UVM testbenches, and
development has been quite rushed.
In case of failed tests, there is quite possibly something wrong with
the agent and not with the RTL.
In any case, a more experienced engineer should probably take a
thorough look at this before merging the PR.

Signed-off-by: ganoam <gnoam@live.com>