Add parallelism to instr_gen_run and iss_sim steps
This commmit enables parallelism of more steps of the dv process by capturing
the commands generated by the riscv-dv tool, then populating and executing a
sub-makefile. This allows us to have parallelism at the level of the makefile
jobserver.
Seperated instr_gen build and run steps

A python script construct_makefile.py is used to capture the raw output commands
and construct a trivial makefile for each stage.

This commit also renames some of the helper variables and stamps for consistency.