Fix spurious rebuilds from dumped vars
Commit f49f452f2 greatly tidied up the file, which is nice, but also
introduced some whitespace around the first argument to vars-prereq.
In Make, this causes chaos!

In particular, we ended up checking whether a variable called
something like "$(last-   gen  -vars-loaded)" was defined. It isn't,
but $(last-gen-vars-loaded) is.

Call strip in the vars-prereq function to get rid of the whitespace at
the "entry point" for all this machinery.