[rtl] Fix MISA
Initially the 'B' bit was going to be used to indicate the presence of
the bit-manipulation extension. Now that has been seperated into
multiple smaller extensions the 'B' MISA bit isn't being used and
remains a reserved bit that should read as 0.

For the RV32B configurations where we implement non ratified
bit-manipulation extensions we must set the 'X' bit.

These changes matche the behaviour of spike.