Restructure ibex_tracer.sv for different versions of Verilator
Recent versions of Verilator complain about the existing code because
it updates decoded_str and data_accessed in functions, and Verilator
doesn't trace the call graph far enough to realise that these are only
triggered from a single process. It then warns that unpredictable
scheduling of always_comb makes this hard to predict.

To avoid the warning, we schedule explicitly! The trick is to update
the two variables with non-blocking assignments on the previous
negedge.

This works... sort of. Unfortunately, the older version of Verilator
that we use in CI was upset with what I originally wrote because it
(I think incorrectly) thought I was mixing blocking and non-blocking
assignments to the file_handle variable.

Gah! This version skirts around that problem by deleting the final
block that was calling $fclose() to cleanly close the file handle. As
far as I can tell, this happens at the end of simulation anyway so
there's no real need to keep it.

I suspect this isn't quite true if the Verilator simulation is running
inside some other process. In that case, maybe the "I'm finishing a
simulation" logic might not call the equivalent of C's atexit(). But I
suspect we'll have other problems too, that aren't caused by this
change, so I'm not too worried :-)