Avoid blocking writes in always_ff blocks in ibex_tracer.sv
This seems a little silly, but Verible's "always_ff_nonblocking" check
tries to ensure that there are no variable updates using blocking
assignments from within an always_ff block. An initialisation as part
of a declaration (like we do for file_name_base) is allowed, but the
update to the "fh" local variable from the previous commit is not.

This seems a little silly to me, but we can cheat here too! We don't
want to open the output file until we have something to write, so we
run the code that opens the file on the preceding negedge and use a
nonblocking assignment to update file_handle, which will have the
right value so that we can write out an instruction on the posedge
that follows.

Bleurgh.