[dv] Add an extra key to common_project_cfg.hjson
The keys in this file get incorporated into the FlowCfg object that
represents a simulation or similar. Adding a spurious key won't cause
any problems, but we actually need it for the next commit, which grabs
the current version of dvsim from OpenTitan. That version of dvsim
expects the "book" key to have been set in common_project_cfg.hjson.

Splitting the two commits like this should make it a bit more obvious
where things have come from.