[dv] Remove phase argument from collect_trans
The prototype of this task has to match the one in dv_base_monitor,
which we are importing from OpenTitan (called "lowrisc_ip").
Unfortunately, OpenTitan imports Ibex, causing a circular reference
which makes it a bit fiddly to change any types.

This commit switches the tasks to match the new prototype we're going
to use in OpenTitan. We can't just apply it in Ibex
immediately (because it won't work with our vendored lowrisc_ip code),
but creating the commit *does* mean we can vendor in the changed Ibex
code at the OpenTitan end.

Once that's sorted, we can vendor OpenTitan back into Ibex and get
everything cleaned up properly.

In hindsight, we probably should have made sure our vendoring
structure was a DAG.