[nix] Add poetry version of ibex python dependencies
This matches the dependencies given in the python-requirements.txt file in the
repo root, but uses the pyproject.toml format and the poetry tool to allow nix
to import this information to construct a python environment.

The poetry.lock file allows each python dep to be pinned, and would need to be
updated using the poetry tool to bump any dependency.

e.g.
```
nix shell nixpkgs#poetry
poetry update
```